projectdeux/pyproject.toml
[tool.poetry]
name = "projectdeux"
version = "0.1.0"
description = "A master thesis project, just a master thesis project, or"
authors = ["Adam S"]
license = "MIT"
readme = "README.md"

# Update this to point to `single_agent`
packages = [
    { include = "agents", from = "src" },
    { include = "integrations", from = "src" },
    { include = "custom_logging", from = "src" },
    { include = "systems", from = "src" },
    { include = "tasks", from = "src" },
    { include = "visualization", from = "src" },
    { include = "entities", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
litellm = "^1.61.13"
python-dotenv = "^1.0.1"
pytest = "^8.3.4"
rerun-sdk = "^0.22.1"
lxml = "^4.9.2"
beautifulsoup4 = "^4.13.3"
requests = "^2.32.3"
selenium = "^4.8.0"  # if you're using Selenium fallback
celery = "^5.4.0"
fastapi = "^0.115.10"
redis = "^5.2.1"
scikit-learn = "^1.6.1"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
